UBW32: Rename files for better conformance to naming conventions
