+++ nuttx/configs/ubw32/src/pic32_leds.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * configs/ubw32/src/up_leds.c
+ * configs/ubw32/src/pic32_leds.c
  *
  *   Copyright (C) 2012, 2015 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
